requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.15.1
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-clearcache==1.2.1
django-crispy-forms==1.10.0
django-filter==2.4.0
django-js-asset==1.2.2
django-social-share==2.1.0
django-widget-tweaks==1.4.8
idna==2.10
oauthlib==3.1.0
Pillow==8.3.2
psycopg2-binary==2.9.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
wincertstore==0.2
